requirements.txt
azure-common==1.1.28
azure-core==1.28.0
azure-functions==1.15.0
azure-identity==1.13.0
azure-keyvault-secrets==4.7.0
certifi==2023.5.7
cffi==1.15.1
charset-normalizer==3.2.0
cryptography==41.0.1
idna==3.4
isodate==0.6.1
msal==1.22.0
msal-extensions==1.0.0
portalocker==2.7.0
pycparser==2.21
PyJWT==2.7.0
requests==2.31.0
six==1.16.0
typing_extensions==4.7.1
urllib3==2.0.3
